+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Upgrade to React 19 and enable the React compiler. In theory this improves UI performance.
+
 ### Added
 - Added Lindormi's Guidance affix.
+- Improve the log button to also offer a "create diagnostics bundle" function.
+
 ### Fixed
 
 ## [7.7.3] - 2026-06-06